requirements.txt
discord.py
async_timeout
PyNaCl
yt-dlp